PREP TO MERGE TO DEVELOP; CLEANED UP CODE; ADDED INTERFACE;
